Moved activity tracker and tracker to data tracker
