+++ src/app/data-tracker/activity-tracker/activity-tracker.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { ActivityService } from '../activity.service';
-import { Exercise } from './exercise.model';
+import { ActivityService } from './activity.service';
+import { Exercise } from '../../shared/models/exercise.model';
 
 
 @Component({
